refactor(app): wave-7 support densify, debug fixtures, size ratchets
Split military/naval scenario Games from host/pump modules, extract shared
debug-handler fixtures, densify panel fixture meta-tests, and add a 600-line
support file-size gate while ratcheting support LOC below 8806.

Refs #4048
