+++ app/test/app_event_handler_debug_spawn_ship_test.dart
@@ -4,38 +4,24 @@ import 'package:colonizethis_app/core/services/debug/app_event_handler_debug_spa
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import 'support/debug_handler_test_fixtures.dart';
+
 void main() {
   suppressLogsForTests();
 
   group('applyDebugShipSpawnAtCapitalHomeFleet', () {
     test('spawns ships into home fleet and advances global ship sequence', () {
-      final game = Game(
+      final game = buildDebugHandlerCapitalGame(
         id: 'g-ship-1',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-          oldWorld: RegionData(
-            provinces: [
-              Province(id: 'oldWorld|1', regionId: 'oldWorld', ownerId: 'p1'),
-            ],
-          ),
-          newWorld: const RegionData(),
-          fleets: [
-            Fleet(
-              id: 'fleet_p1',
-              ownerId: 'p1',
-              inPortAtProvinceId: 'oldWorld|1',
-              regionId: 'oldWorld',
-              ships: const [ShipInstance(id: 'ship_3', typeId: 'carrack')],
-            ),
-          ],
-          nextShipInstanceSeq: 4,
-        ),
-        players: const [
-          Player(
-            id: 'p1',
-            displayName: 'P1',
-            isHuman: true,
-            capitalProvinceId: 'oldWorld|1',
+        includeCapitalTile: false,
+        nextShipInstanceSeq: 4,
+        fleets: [
+          Fleet(
+            id: 'fleet_p1',
+            ownerId: 'p1',
+            inPortAtProvinceId: 'oldWorld|1',
+            regionId: 'oldWorld',
+            ships: const [ShipInstance(id: 'ship_3', typeId: 'carrack')],
           ),
         ],
       );
@@ -57,26 +43,9 @@ void main() {
     });
 
     test('fails when player has no valid home fleet at capital', () {
-      final game = Game(
+      final game = buildDebugHandlerCapitalGame(
         id: 'g-ship-2',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-          oldWorld: RegionData(
-            provinces: [
-              Province(id: 'oldWorld|1', regionId: 'oldWorld', ownerId: 'p1'),
-            ],
-          ),
-          newWorld: const RegionData(),
-          fleets: const [],
-        ),
-        players: const [
-          Player(
-            id: 'p1',
-            displayName: 'P1',
-            isHuman: true,
-            capitalProvinceId: 'oldWorld|1',
-          ),
-        ],
+        includeCapitalTile: false,
       );
       const event = SpawnDebugShipAtCapitalHomeFleetEvent(
         humanPlayerId: 'p1',
